+++ src/main/php/lang/Type.class.php
@@ -166,15 +166,17 @@ class Type extends Object {
     } else if ('void' === $type) {
       return self::$VOID;
     } else if ('array' === $type) {
-      return Type::$ARRAY;
-    } else if (0 === substr_compare($type, '[]', -2)) {
-      return new ArrayType(substr($type, 0, -2));
-    } else if (0 === substr_compare($type, '[:', 0, 2)) {
-      return new MapType(substr($type, 2, -1));
+      return self::$ARRAY;
     } else if (0 === substr_compare($type, 'function(', 0, 9)) {
       return FunctionType::forName($type);
+    } else if (0 === substr_compare($type, '[]', -2)) {
+      return new ArrayType(self::forName(substr($type, 0, -2)));
+    } else if (0 === substr_compare($type, '[:', 0, 2)) {
+      return new MapType(self::forName(substr($type, 2, -1)));
+    } else if (0 === substr_compare($type, '(function(', 0, 10)) {
+      return FunctionType::forName(substr($type, 1, -1));
     } else if (0 === substr_compare($type, '*', -1)) {
-      return new ArrayType(substr($type, 0, -1));
+      return new ArrayType(self::forName(substr($type, 0, -1)));
     } else if (false === ($p= strpos($type, '<'))) {
       $normalized= strtr($type, '\\', '.');
       return strstr($normalized, '.') ? XPClass::forName($normalized) : new XPClass($normalized);